requirements.txt
Flask>=2.3
reportlab>=4.0
Pillow>=10.0
